chore(@clayui/link): update Link component stories
